pyproject.toml
[tool.poetry]
name = "bnn-analysis"
version = "0.1.0"
description = "Performance analysis of bits neural networks and binarized neural networks"
authors = ["nickolasrm"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
hydra-core = "^1.3.1"
gym = "^0.26.2"
numpy = ">=1.23.0,<1.24.0"
pandas = "^1.5.3"
seaborn = "^0.12.2"
tensorflow = "^2.11.0"
pygad = "^2.18.1"
larq = "^0.13.0"
datasets = "^2.9.0"
scikit-learn = "^1.2.1"
wandb = "^0.13.9"
click = "^8.1.3"
keras = "^2.11.0"
flatten-dict = "^0.4.2"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.5.3"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pylint = "^2.15.10"
mypy = "^0.991"
pydocstyle = "^6.3.0"
black = {version = "^23.1a1", allow-prereleases = true}
flake8 = "^6.0.0"
pre-commit = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
